Moved all screens to pages
