+++ src/app/routes/gifts/details.tsx
@@ -1,9 +1,19 @@
-import { type FC } from 'react';
+import type { WebApp } from '@twa-dev/types';
+
+import { type FC, useEffect, useState } from 'react';
+import { useNavigate, useSearchParams } from 'react-router-dom';
 
 import { json } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 
-import { GiftDetailsPage } from '@/screens/gifts';
+import { $path } from 'remix-routes';
+
+import { useTelegram } from '@/common/providers';
+import { createTgLink } from '@/common/telegram';
+import type { TgLink } from '@/common/telegram/types.ts';
+import { useBackButton } from '@/common/telegram/useBackButton.ts';
+import { useMainButton } from '@/common/telegram/useMainButton.ts';
+import { GiftDetails, Icon } from '@/components';
 
 export const loader = () => {
   return json({
@@ -13,9 +23,39 @@ export const loader = () => {
 };
 
 const Page: FC = () => {
-  const data = useLoaderData<typeof loader>();
+  const { botUrl, appName } = useLoaderData<typeof loader>();
 
-  return <GiftDetailsPage botUrl={data.botUrl} appName={data.appName} />;
+  const navigate = useNavigate();
+  const [searchParams] = useSearchParams();
+  const { webApp } = useTelegram();
+  const { addBackButton } = useBackButton();
+  const { mainButton } = useMainButton();
+
+  const [link, setLink] = useState<TgLink | null>(null);
+
+  useEffect(() => {
+    mainButton.show();
+    addBackButton(() => {
+      navigate($path('/gifts'));
+    });
+
+    setLink(
+      createTgLink({
+        secret: searchParams.get('seed') as string,
+        symbol: searchParams.get('symbol') as string,
+        amount: searchParams.get('balance') as string,
+        botUrl,
+        appName,
+      }),
+    );
+  }, []);
+
+  return (
+    <div className="grid items-center justify-center h-[93vh]">
+      <Icon name="Present" size={250} className="justify-self-center mt-auto" />
+      <GiftDetails link={link} webApp={webApp as WebApp} />
+    </div>
+  );
 };
 
 export default Page;
