--- src/screens/gifts/index.ts
@@ -1,4 +1,0 @@
-import GiftDetailsPage from './GiftDetailsPage';
-import GiftPage from './GiftPage';
-
-export { GiftDetailsPage, GiftPage };
