+++ src/screens/onboarding/index.ts
@@ -1,8 +1,2 @@
-import CreateWalletPage from './CreateWallet';
-import OnboardingPage from './Onboarding';
-import PasswordPage from './Password';
-
 export { OnboardingStartPage } from './OnboardingStartPage';
 export { RestoreWalletPage } from './RestoreWallet';
-
-export { CreateWalletPage, PasswordPage, OnboardingPage };
