--- src/screens/settings/index.ts
@@ -1,15 +1,0 @@
-import SettingsBackupPage from './BackupPage';
-import RecoveryPage from './RecoveryPage';
-import SettingsPage from './SettingsPage';
-import ChangePasswordPage from './password/ChangePassword';
-import NewPasswordPage from './password/NewPassword';
-import PasswordConfirmationPage from './password/PasswordConfirmation';
-
-export {
-  SettingsBackupPage,
-  SettingsPage,
-  ChangePasswordPage,
-  NewPasswordPage,
-  PasswordConfirmationPage,
-  RecoveryPage,
-};
